backend/requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
PyPDF2
python-multipart